KY/aug_stage/stage1_survey_persona_aug/requirements.txt

python-dotenv>=1.0.0
pydantic>=2.0.0
datasets>=2.0.0
tqdm>=4.0.0
openai>=1.0.0
anthropic>=0.18.0
